Remove ingredient refill

+++ Assets/Scripts/ScriptableObjects/Worker.cs
@@ -6,7 +6,7 @@ public class Worker : ScriptableObject
     public string id;
     public int level = 1;
     public float secondsBetweenTasks;
-    public int pricePerDay;
+    public float pricePerDay;
     public WorkersRole role = WorkersRole.GRILL;
 
 }
