+++ Assets/Scripts/Visuals/WalletVisual.cs
@@ -5,7 +5,7 @@ public class WalletVisual : MonoBehaviour
 {
     [SerializeField] private TMP_Text walletText;
 
-    public void UpdateWalletAmount(int amount)
+    public void UpdateWalletAmount(float amount)
     {
         walletText.text = amount.ToString() + " €";
     }
